+++ src/mcp_orch/models/__init__.py
@@ -11,6 +11,7 @@ from .server_log import ServerLog, LogLevel, LogCategory
 from .favorite import UserFavorite
 from .client_session import ClientSession
 from .tool_call_log import ToolCallLog, CallStatus
+from .tool_preference import ToolPreference
 from .worker_config import WorkerConfig
 from .activity import Activity, ActivityType, ActivitySeverity, ProjectActivity
 
@@ -37,6 +38,7 @@ __all__ = [
     "ClientSession",
     "ToolCallLog",
     "CallStatus",
+    "ToolPreference",
     "WorkerConfig",
     "Activity",
     "ProjectActivity",